Get dtype without using .data.

+++ improver/categorical/modal_code.py
@@ -320,7 +320,7 @@ class ModalCategory(BaseModalCategory):
             result = cube
         else:
             result = cube.collapsed("time", self.aggregator_instance)
-            result.data = result.data.astype(cube.data.dtype)
+            result.data = result.data.astype(cube.dtype)
         self._set_blended_times(result)
 
         result = self._prepare_result_cube(
